pyproject.toml
[project]
name = "ai-dial-general-purpose-agent-with-memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aidial-client==0.3.0",
    "aidial-sdk==0.27.0",
    "beautifulsoup4==4.14.2",
    "faiss-cpu==1.12.0",
    "langchain==1.0.3",
    "langchain-text-splitters==1.0.0",
    "mcp==1.20.0",
    "numpy==2.3.4",
    "pandas==2.3.3",
    "pdfplumber==0.11.7",
    "pydantic==2.12.3",
    "sentence-transformers==5.1.2",
    "tabulate==0.9.0",
]

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "task"
module-root = ""
